Extract dropdown components and repair badge
